feat: refactor new feed to component
